refactor: move token repo to session aggregate
