+++ packages/server/src/aggregates/session/domain/index.ts
@@ -1,1 +1,2 @@
 export * from './session.abstract.service';
+export * from './repositories';
