+++ packages/server/src/aggregates/session/domain/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './token.abstract.repository';
