+++ packages/server/src/aggregates/session/infrastructure/index.ts
@@ -1,3 +1,5 @@
 export * from './session.module';
 export * from './session.service';
 export * from './session.unit-of-work';
+export * from './entities';
+export * from './repositories';
