+++ packages/server/src/aggregates/session/infrastructure/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './token.repository';
