+++ packages/server/src/aggregates/session/infrastructure/session.unit-of-work.ts
@@ -4,12 +4,15 @@ import { DataSource, EntityManager } from 'typeorm';
 import { TypeormUnitOfWork } from '@/core/unit-of-work';
 import { SessionAbstractRepository } from '@/entities/session/domain';
 import { SessionRepository } from '@/entities/session/infrastructure';
+import { TokenAbstractRepository } from '../domain';
+import { TokenRepository } from './repositories';
 import { UserAbstractRepository } from '@/entities/user/domain';
 import { UserRepository } from '@/entities/user/infrastructure';
 
 export interface SessionUnitOfWorkContext {
   sessionRepository: SessionAbstractRepository;
   userRepository: UserAbstractRepository;
+  tokenRepository: TokenAbstractRepository;
 }
 
 @Injectable()
@@ -24,10 +27,12 @@ export class SessionUnitOfWork extends TypeormUnitOfWork<SessionUnitOfWorkContex
       entityManager,
       userRepository,
     );
+    const tokenRepository = new TokenRepository(entityManager, sessionRepository);
 
     return {
       sessionRepository,
       userRepository,
+      tokenRepository,
     };
   }
 }
