+++ packages/server/src/aggregates/token-auth/domain/repositories/index.ts
@@ -1,2 +1,1 @@
 export * from './password.abstract.repository';
-export * from './token.abstract.repository';
