+++ packages/server/src/aggregates/token-auth/infrastructure/entities/index.ts
@@ -1,2 +1,1 @@
 export * from './password.repository.entity';
-export * from './token.repository.entity';
