+++ packages/server/src/aggregates/token-auth/infrastructure/repositories/index.ts
@@ -1,2 +1,1 @@
 export * from './password.repository';
-export * from './token.repository';
