+++ packages/server/src/aggregates/token-auth/infrastructure/token-auth.unit-of-work.ts
@@ -1,23 +1,19 @@
 import { Injectable } from '@nestjs/common';
 import { DataSource, EntityManager } from 'typeorm';
 
-import {
-  PasswordAbstractRepository,
-  TokenAbstractRepository,
-} from '@/aggregates/token-auth/domain';
+import { PasswordAbstractRepository } from '@/aggregates/token-auth/domain';
 import { TypeormUnitOfWork } from '@/core/unit-of-work';
 import { SessionAbstractRepository } from '@/entities/session/domain';
 import { SessionRepository } from '@/entities/session/infrastructure';
 import { UserAbstractRepository } from '@/entities/user/domain';
 import { UserRepository } from '@/entities/user/infrastructure';
 
-import { PasswordRepository, TokenRepository } from './repositories';
+import { PasswordRepository } from './repositories';
 
 export interface TokenAuthUnitOfWorkContext {
   userRepository: UserAbstractRepository;
   passwordRepository: PasswordAbstractRepository;
   sessionRepository: SessionAbstractRepository;
-  tokenRepository: TokenAbstractRepository;
 }
 
 @Injectable()
@@ -37,7 +33,6 @@ export class TokenAuthUnitOfWork extends TypeormUnitOfWork<TokenAuthUnitOfWorkCo
       userRepository,
       passwordRepository: new PasswordRepository(entityManager, userRepository),
       sessionRepository,
-      tokenRepository: new TokenRepository(entityManager, sessionRepository),
     };
   }
 }
